Update oe_configure_layer.py
TWMS and WMTS files are created even if the no_twms and no_wmts parameters are provided.  Ideally, if we provide those parameters, we would not see any configuration files created.  The nuance is that we need a TWMS file is needed for the WMTS.  So I'm just creating a temporary file that can be written/read as though it's the actual TWMS mrf file.