GITC-1217 Fixed failing WMS CI tests
One difference to note between OnEarth 1.x and 2.x:
Since we're using WMTS as the backing for WMS, we get blank tiles when
out-of-range date/times are requested. In OnEarth 1.x, MapServer would
return a 400 error due to the MRF file not actually existing on disk.

Commented out WMS time format validation tests. This has not been
implemented yet in OnEarth 2.x. A filter module is likely needed to
support this.