Cherrypicking NearNB vs NNB update from 1.4.0
In GDAL 2.4.x, we need to provide `Avg` or `NearNB` to mrf_insert, while we can provide `Avg` or `NNB` to gdaladdo. In GDAL 3.1.x (once Lucian merges), we will be able to provide `Avg` or `NNB` to both. Note, that there is no case sensitivity on any of those values.