GITC-2114 Use the last GC ResourceURL when configuring reproject layers.
This is a bug introduced by the addition of multiple GetCapabilities
ResourceURLs in GITC-1297. Reproject is using only the first ResourceURL
that it finds, but it needs to use the last one, which has {Time}
included in the URL string. oe2_reproject_configure is updated to now
use the last one.